+++ py_basic_commands/json_scripts/__init__.py
@@ -2,3 +2,4 @@ from py_basic_commands.json_scripts.create_json   import create_json, CreateJson
 from py_basic_commands.json_scripts.read_json     import read_json, ReadJson
 from py_basic_commands.json_scripts.write_json    import write_json, WriteJson
 from py_basic_commands.json_scripts.json_editor   import JsonEditor
+from py_basic_commands.json_scripts.prettify_json      import prettify_json